requirements.txt
discord.py
flask
flask-basicauth
requests
